requirements.txt
Pillow~=8.0.1
flux_led~=0.22
